+++ src/pages/offrir/Index.jsx
@@ -4,7 +4,7 @@ import Table from "../../components/ui/Table";
 import Message from "../../components/ui/Message";
 import API from "../../services/api";
 import Loading from "../../components/ui/Loading";
-import HideMessage from "../../components/ui/hideMessage";
+import MessageAlert from "../../components/ui/MessageAlert";
 
 const IndexOffrir = () => {
   const [offres, setOffres] = useState([]);
@@ -40,6 +40,8 @@ const IndexOffrir = () => {
   }, []);
 
   const handleDelete = async (item) => {
+    const confirmed = await MessageAlert(`la filiere "${item.filiere_info}"`);
+    if (!confirmed) return;
     try {
       await API.delete(`/offres-formations/${item.id}`);
       setOffres((prev) => prev.filter((o) => o.id !== item.id));
@@ -68,16 +70,16 @@ const IndexOffrir = () => {
   return (
     <div className="space-y-6">
       <div className="flex justify-between items-center">
-        <h1 className="text-2xl font-bold">Liste des offres de formation</h1>
+        <h1 className="text-2xl font-bold">Liste des filieres offrés par l'établissement</h1>
         <button
           onClick={() => navigate("/offres-formations/create")}
           className="px-4 py-2 rounded-xl bg-gradient-to-r from-indigo-600 to-purple-400 text-white font-semibold shadow-md hover:opacity-90 transition"
         >
-          + Ajouter une offre
+          + Ajouter une filiere
         </button>
       </div>
       {message && <Message type={message.type} text={message.text} />}
-      <HideMessage message={message} onHide={() => setMessage(null)} />
+      
       <Table
         columns={columns}
         data={offres}
